move some builder interatiction  behaviors into Analytics::Export::Buildable, initial build out for a Buildable spec
