+++ lib/analytics/export.rb
@@ -1,5 +1,6 @@
 require_relative "export/model"
 require_relative "export/builder"
+require_relative "export/buildable"
 
 module Analytics
   module Export
