+++ spec/models/course_analytics_export_spec.rb
@@ -17,6 +17,10 @@ describe CourseAnalyticsExport do
     expect(subject).to respond_to :object_key_microseconds
   end
 
+  it "includes Analytics::Export::Buildable" do
+    expect(subject).to respond_to :upload_builder_archive_to_s3
+  end
+
   describe "#generate_secure_token" do
     it "creates a new secure token with the export data" do
       token = subject.generate_secure_token
@@ -70,37 +74,20 @@ describe CourseAnalyticsExport do
     end
   end
 
-  describe "#build_archive!" do
-    it "calls #build_archive! on the export builder" do
-      builder = double :builder, build_archive!: "..going.."
-      allow(subject).to receive(:export_builder) { builder }
-      expect(subject.build_archive!).to eq "..going.."
-    end
-  end
-
-  describe "#export_builder" do
-    before do
+  describe "#export_builder_attrs" do
+    it "returns a hash of attributes to use for the builder" do
       allow(subject).to receive_messages \
         export_context: double(:context, export_data: "some data"),
         export_classes: "some classes",
         url_safe_filename: "the_filename.txt",
         formatted_course_number: "ECO500"
-    end
 
-    it "builds a new export builder" do
-      builder = subject.export_builder
-      expect(builder.class).to eq Analytics::Export::Builder
-      expect(builder.export_data).to eq "some data"
-      expect(builder.export_classes).to eq "some classes"
-      expect(builder.filename).to eq "the_filename.txt"
-      expect(builder.directory_name).to eq "ECO500"
-    end
-
-    it "caches the builder" do
-      subject.export_builder
-      expect(Analytics::Export::Builder).not_to receive(:new)
-      expect(subject.instance_variable_get :@export_builder)
-        .to eq subject.export_builder
+      expect(subject.export_builder_attrs).to eq({
+        export_data: "some data",
+        export_classes: "some classes",
+        filename: "the_filename.txt",
+        directory_name: "ECO500"
+      })
     end
   end
